requirements.txt
openai==0.28.0
gradio
tiktoken
langchain
langchain-openai
langchain-community
chromadb
python-dotenv